orchestrator/requirements.txt
fastapi==0.115.12
pyyaml==6.0.2
uvicorn==0.34.3
httpx==0.28.1
pydantic==2.11.4
# Docker SDK — used by neo4j_manager to stop/restart the neo4j container
docker==7.1.0
# Token counting for compaction threshold and Pass 1 activation
tiktoken==0.9.0
# MCP client for Pass 1 substrate tool calls
mcp==1.9.4
# SSE streaming support
sse-starlette==2.1.3
# Testing
pytest==8.3.5
pytest-asyncio==0.25.3
respx==0.21.1
